+++ packages/celotool/src/lib/helm_deploy.ts
@@ -15,14 +15,7 @@ import {
   spawnCmd,
   spawnCmdWithExitOnFailure,
 } from './cmd-utils'
-import {
-  EnvTypes,
-  envVar,
-  fetchEnv,
-  fetchEnvOrFallback,
-  isProduction,
-  monorepoRoot,
-} from './env-utils'
+import { EnvTypes, envVar, fetchEnv, fetchEnvOrFallback, monorepoRoot } from './env-utils'
 import { ensureAuthenticatedGcloudAccount } from './gcloud_utils'
 import { generateGenesisFromEnv } from './generate_utils'
 import {
@@ -47,6 +40,7 @@ const BACKUP_GCS_SECRET_NAME = 'backup-blockchain-credentials'
 const TIMEOUT_FOR_LOAD_BALANCER_POLL = 1000 * 60 * 25 // 25 minutes
 const LOAD_BALANCER_POLL_INTERVAL = 1000 * 10 // 10 seconds
 
+const TESTNET_CHART_DIR = '../helm-charts/testnet'
 export type HelmAction = 'install' | 'upgrade'
 
 async function validateExistingCloudSQLInstance(instanceName: string) {
@@ -764,7 +758,7 @@ async function helmIPParameters(celoEnv: string) {
 
 async function helmParameters(celoEnv: string, useExistingGenesis: boolean) {
   const valueFilePath = `/tmp/${celoEnv}-testnet-values.yaml`
-  await saveHelmValuesFile(celoEnv, valueFilePath, useExistingGenesis)
+  await saveHelmValuesFile(celoEnv, valueFilePath, useExistingGenesis, false)
 
   const gethMetricsOverrides =
     fetchEnvOrFallback('GETH_ENABLE_METRICS', 'false') === 'true'
@@ -816,11 +810,6 @@ async function helmParameters(celoEnv: string, useExistingGenesis: boolean) {
     `--set geth.gstorage_data_bucket=${fetchEnvOrFallback('GSTORAGE_DATA_BUCKET', '')}`,
     `--set geth.faultyValidators="${fetchEnvOrFallback('FAULTY_VALIDATORS', '0')}"`,
     `--set geth.faultyValidatorType="${fetchEnvOrFallback('FAULTY_VALIDATOR_TYPE', '0')}"`,
-    // Disable by default block age check in fullnode readinessProbe except for production envs
-    `--set geth.fullnodeCheckBlockAge=${fetchEnvOrFallback(
-      envVar.FULL_NODE_READINESS_CHECK_BLOCK_AGE,
-      `${isProduction()}`
-    )}`,
     `--set geth.tx_nodes="${fetchEnv('TX_NODES')}"`,
     `--set geth.private_tx_nodes="${fetchEnv(envVar.PRIVATE_TX_NODES)}"`,
     `--set geth.ssd_disks="${fetchEnvOrFallback(envVar.GETH_NODES_SSD_DISKS, 'true')}"`,
@@ -842,10 +831,9 @@ async function helmParameters(celoEnv: string, useExistingGenesis: boolean) {
 }
 
 async function helmCommand(command: string, pipeOutput = false) {
-  if (isCelotoolVerbose() && !command.includes(' dep build ')) {
-    await execCmdWithExitOnFailure(command + ' --dry-run --debug')
-  } else if (isCelotoolVerbose()) {
-    await execCmdWithExitOnFailure(command + ' --debug')
+  // "helm diff" is a plugin and doesn't support "--debug"
+  if (isCelotoolVerbose() && !command.startsWith('helm diff')) {
+    command += ' --debug'
   }
 
   await execCmdWithExitOnFailure(command, {}, pipeOutput)
@@ -994,21 +982,32 @@ export async function removeGenericHelmChart(releaseName: string, namespace: str
   }
 }
 
+function getExtraValuesFile(celoEnv: string) {
+  const extraValuesFile = fs.existsSync(`${TESTNET_CHART_DIR}/values-${celoEnv}.yaml`)
+    ? `values-${celoEnv}.yaml`
+    : undefined
+  return extraValuesFile
+}
+
 export async function installHelmChart(celoEnv: string, useExistingGenesis: boolean) {
   await failIfSecretMissing(BACKUP_GCS_SECRET_NAME, 'default')
   await copySecret(BACKUP_GCS_SECRET_NAME, 'default', celoEnv)
+  const extraValuesFile = getExtraValuesFile(celoEnv)
   return installGenericHelmChart(
     celoEnv,
     celoEnv,
-    '../helm-charts/testnet',
-    await helmParameters(celoEnv, useExistingGenesis)
+    TESTNET_CHART_DIR,
+    await helmParameters(celoEnv, useExistingGenesis),
+    true,
+    extraValuesFile
   )
 }
 
 export async function upgradeHelmChart(celoEnv: string, useExistingGenesis: boolean) {
   console.info(`Upgrading helm release ${celoEnv}`)
   const parameters = await helmParameters(celoEnv, useExistingGenesis)
-  await upgradeGenericHelmChart(celoEnv, celoEnv, '../helm-charts/testnet', parameters)
+  const extraValuesFile = getExtraValuesFile(celoEnv)
+  await upgradeGenericHelmChart(celoEnv, celoEnv, TESTNET_CHART_DIR, parameters, extraValuesFile)
 }
 
 export async function resetAndUpgradeHelmChart(celoEnv: string, useExistingGenesis: boolean) {
@@ -1197,7 +1196,10 @@ async function generateMyCeloGenesis(): Promise<string> {
   // Clean up the tmp dir
   await spawnCmd('rm', ['-rf', celoBlockchainDir], { silent: true })
   fs.mkdirSync(celoBlockchainDir)
-  const gethTag = fetchEnv(envVar.GETH_NODE_DOCKER_IMAGE_TAG)
+  const gethTag =
+    fetchEnvOrFallback(envVar.GETH_MYCELO_COMMIT, '') !== ''
+      ? fetchEnv(envVar.GETH_MYCELO_COMMIT)
+      : fetchEnv(envVar.GETH_NODE_DOCKER_IMAGE_TAG)
   const celoBlockchainVersion = gethTag.includes('.') ? `v${gethTag}` : gethTag
   await checkoutGethRepo(celoBlockchainVersion, celoBlockchainDir)
   await buildGethAll(celoBlockchainDir)
@@ -1242,8 +1244,8 @@ async function generateMyCeloGenesis(): Promise<string> {
   })
   const genesisPath = path.join(celoBlockchainDir, 'genesis.json')
   const genesisContent = fs.readFileSync(genesisPath).toString()
-  return genesisContent
 
   // Clean up the tmp dir as it's no longer needed
   await spawnCmd('rm', ['-rf', celoBlockchainDir], { silent: true })
+  return genesisContent
 }