+++ packages/phone-number-privacy/signer/src/config.ts
@@ -19,6 +19,7 @@ export enum SupportedDatabase {
   Postgres = 'postgres', // PostgresSQL
   MySql = 'mysql', // MySQL
   MsSql = 'mssql', // Microsoft SQL Server
+  Sqlite = 'sqlite3', // SQLite (for testing)
 }
 
 export enum SupportedKeystore {
@@ -47,6 +48,7 @@ interface Config {
   }
   blockchain: {
     provider: string
+    apiKey?: string
   }
   db: {
     type: SupportedDatabase
@@ -105,6 +107,7 @@ const config: Config = {
   },
   blockchain: {
     provider: env.BLOCKCHAIN_PROVIDER,
+    apiKey: env.BLOCKCHAIN_API_KEY,
   },
   db: {
     type: env.DB_TYPE ? env.DB_TYPE.toLowerCase() : SupportedDatabase.Postgres,