+++ packages/phone-number-privacy/signer/src/database/models/domainState.ts
@@ -1,5 +1,9 @@
-import { Domain, domainHash, isSequentialDelayDomain } from '@celo/identity/lib/odis/domains'
 import { SequentialDelayDomainState, WarningMessage } from '@celo/phone-number-privacy-common'
+import {
+  Domain,
+  domainHash,
+  isSequentialDelayDomain,
+} from '@celo/phone-number-privacy-common/lib/domains'
 
 export const DOMAINS_STATES_TABLE = 'domainsStates'
 export enum DOMAINS_STATES_COLUMNS {
@@ -8,8 +12,8 @@ export enum DOMAINS_STATES_COLUMNS {
   timer = 'timer',
   disabled = 'disabled',
 }
-export class DomainState {
-  public static createEmptyDomainState(domain: Domain): DomainState {
+export class DomainStateRecord {
+  public static createEmptyDomainState(domain: Domain): DomainStateRecord {
     if (isSequentialDelayDomain(domain)) {
       return {
         [DOMAINS_STATES_COLUMNS.domainHash]: domainHash(domain).toString('hex'),
@@ -19,6 +23,11 @@ export class DomainState {
       }
     }
 
+    // canary provides a compile-time check that all subtypes of Domain have branches. If a case
+    // was missed, then an error will report that domain cannot be assigned to type `never`.
+    const canary = (x: never) => x
+    canary(domain)
+
     throw new Error(WarningMessage.UNKNOWN_DOMAIN)
   }
 
@@ -27,10 +36,10 @@ export class DomainState {
   [DOMAINS_STATES_COLUMNS.timer]: number | undefined;
   [DOMAINS_STATES_COLUMNS.disabled]: boolean
 
-  constructor(domainState: SequentialDelayDomainState) {
-    this[DOMAINS_STATES_COLUMNS.domainHash] = domainState[DOMAINS_STATES_COLUMNS.domainHash]
-    this[DOMAINS_STATES_COLUMNS.counter] = domainState[DOMAINS_STATES_COLUMNS.counter]
-    this[DOMAINS_STATES_COLUMNS.timer] = domainState[DOMAINS_STATES_COLUMNS.timer]
-    this[DOMAINS_STATES_COLUMNS.disabled] = domainState[DOMAINS_STATES_COLUMNS.disabled]
+  constructor(hash: string, domainState: SequentialDelayDomainState) {
+    this[DOMAINS_STATES_COLUMNS.domainHash] = hash
+    this[DOMAINS_STATES_COLUMNS.counter] = domainState.counter
+    this[DOMAINS_STATES_COLUMNS.timer] = domainState.timer
+    this[DOMAINS_STATES_COLUMNS.disabled] = domainState.disabled
   }
 }