+++ packages/sdk/identity/src/odis/query.ts
@@ -1,14 +1,26 @@
-// Utilities for interacting with the Oblivious Decentralized Identifier Service (ODIS)
-
-import { hexToBuffer, trimLeading0x } from '@celo/base/lib/address'
+import { hexToBuffer } from '@celo/base/lib/address'
 import { selectiveRetryAsyncWithBackOff } from '@celo/base/lib/async'
 import { ContractKit } from '@celo/contractkit'
+import {
+  AuthenticationMethod,
+  CombinerEndpoint,
+  Domain,
+  DomainEndpoint,
+  DomainRequest,
+  DomainResponse,
+  GetBlindedMessageSigRequest,
+  GetContactMatchesRequest,
+  GetContactMatchesResponse,
+  PhoneNumberPrivacyEndpoint,
+  PhoneNumberPrivacyRequest,
+} from '@celo/phone-number-privacy-common'
 import fetch from 'cross-fetch'
+import crypto from 'crypto'
 import debugFactory from 'debug'
-import { ec as EC } from 'elliptic'
+import { isLeft } from 'fp-ts/lib/Either'
+import * as t from 'io-ts'
 
 const debug = debugFactory('kit:odis:query')
-const ec = new EC('secp256k1')
 
 export interface WalletKeySigner {
   authenticationMethod: AuthenticationMethod.WALLET_KEY
@@ -20,58 +32,32 @@ export interface EncryptionKeySigner {
   rawKey: string
 }
 
-export interface CustomSigner {
-  authenticationMethod: AuthenticationMethod.CUSTOM_SIGNER
-  customSigner: (body: string) => Promise<string>
-}
-
 // Support signing with the DEK or with the
-export type AuthSigner = WalletKeySigner | EncryptionKeySigner | CustomSigner
+export type AuthSigner = WalletKeySigner | EncryptionKeySigner
 
-export enum AuthenticationMethod {
-  WALLET_KEY = 'wallet_key',
-  ENCRYPTION_KEY = 'encryption_key',
-  CUSTOM_SIGNER = 'custom_signer',
-}
+// Re-export types and aliases to maintain backwards compatibility.
+export { AuthenticationMethod, PhoneNumberPrivacyRequest }
+export type SignMessageRequest = GetBlindedMessageSigRequest
+export type MatchmakingRequest = GetContactMatchesRequest
+export type MatchmakingResponse = GetContactMatchesResponse
 
-// TODO(victor) Requests here are duplicated in. They should be deduplicated.
-// https://github.com/celo-org/celo-monorepo/blob/d5a275b56ca62360d1da9d00d38870888c9dbada/packages/phone-number-privacy/common/src/interfaces/requests.ts#L4
-export interface PhoneNumberPrivacyRequest {
-  account: string
-  authenticationMethod: AuthenticationMethod
-  version?: string
-  sessionID?: string
-}
-
-export interface SignMessageRequest extends PhoneNumberPrivacyRequest {
-  blindedQueryPhoneNumber: string
-  hashedPhoneNumber?: string
-}
-
-export interface MatchmakingRequest extends PhoneNumberPrivacyRequest {
-  userPhoneNumber: string
-  contactPhoneNumbers: string[]
-  hashedPhoneNumber: string
-  signedUserPhoneNumber?: string
-}
-
-export interface SignMessageResponse {
+// Combiner returns a response inconsistent with the SignMessageResponse defined in
+// @celo/phone-number-privacy-common. Combiner response type is defined here as a result.
+export interface CombinerSignMessageResponse {
   success: boolean
   combinedSignature: string
 }
-
-export interface MatchmakingResponse {
-  success: boolean
-  matchedContacts: Array<{
-    phoneNumber: string
-  }>
-}
+/** @deprecated Exported as SignMessageResponse for backwards compatibility. */
+export type SignMessageResponse = CombinerSignMessageResponse
 
 export enum ErrorMessages {
   ODIS_QUOTA_ERROR = 'odisQuotaError',
+  ODIS_RATE_LIMIT_ERROR = 'odisRateLimitError',
   ODIS_INPUT_ERROR = 'odisBadInputError',
   ODIS_AUTH_ERROR = 'odisAuthError',
   ODIS_CLIENT_ERROR = 'Unknown Client Error',
+  ODIS_FETCH_ERROR = 'odisFetchError',
+  ODIS_RESPONSE_ERROR = 'odisResponseError',
 }
 
 export interface ServiceContext {
@@ -108,49 +94,57 @@ export function getServiceContext(contextName = 'mainnet') {
   }
 }
 
-export function signWithDEK(message: string, signer: EncryptionKeySigner) {
+export function signWithDEK(msg: string, signer: EncryptionKeySigner) {
+  return signWithRawKey(msg, signer.rawKey)
+}
+
+export function signWithRawKey(msg: string, rawKey: string) {
+  // NOTE: Elliptic will truncate the raw msg to 64 bytes before signing,
+  // so make sure to always pass the hex encoded msgDigest instead.
+  const msgDigest = crypto.createHash('sha256').update(JSON.stringify(msg)).digest('hex')
+
+  // NOTE: elliptic is disabled elsewhere in this library to prevent
+  // accidental signing of truncated messages.
+  // tslint:disable-next-line:import-blacklist
+  const EC = require('elliptic').ec
+  const ec = new EC('secp256k1')
+
   // Sign
-  const key = ec.keyFromPrivate(hexToBuffer(signer.rawKey))
-  const sig = JSON.stringify(key.sign(message).toDER())
-  // Verify
-  const dek = key.getPublic(true, 'hex')
-  const pubkey = ec.keyFromPublic(trimLeading0x(dek), 'hex')
-  const validSignature: boolean = pubkey.verify(message, JSON.parse(sig))
-  debug(`Signature is valid: ${validSignature} signed by ${dek}`)
-  return sig
+  const key = ec.keyFromPrivate(hexToBuffer(rawKey))
+  return JSON.stringify(key.sign(msgDigest).toDER())
 }
 
 /**
  * Make a request to lookup the phone number identifier or perform matchmaking
- * @param signer type of key to sign with
- * @param body request body
- * @param context contains service URL
- * @param endpoint endpoint to hit
+ * @param signer Type of key to sign with. May be undefined if the request is presigned.
+ * @param body Request to send in the body of the HTTP request.
+ * @param context Contains service URL and public to determine which instance to contact.
+ * @param endpoint Endpoint to query (e.g. '/getBlindedMessagePartialSig', '/getContactMatches').
  */
 export async function queryOdis<ResponseType>(
   signer: AuthSigner,
   body: PhoneNumberPrivacyRequest,
   context: ServiceContext,
-  endpoint: string
-) {
+  endpoint: PhoneNumberPrivacyEndpoint | CombinerEndpoint
+): Promise<ResponseType> {
   debug(`Posting to ${endpoint}`)
 
-  // Sign payload using account privkey
   const bodyString = JSON.stringify(body)
 
-  let authHeader = ''
+  // Sign payload using provided account and authentication method.
+  let signature: string
   if (signer.authenticationMethod === AuthenticationMethod.ENCRYPTION_KEY) {
-    authHeader = signWithDEK(bodyString, signer as EncryptionKeySigner)
+    signature = signWithDEK(bodyString, signer as EncryptionKeySigner)
   } else if (signer.authenticationMethod === AuthenticationMethod.WALLET_KEY) {
-    authHeader = await signer.contractKit.connection.sign(bodyString, body.account)
-  } else {
-    authHeader = await signer.customSigner(bodyString)
+    const account = body.account
+    signature = await signer.contractKit.connection.sign(bodyString, account)
   }
 
   const { odisUrl } = context
 
   const dontRetry = [
     ErrorMessages.ODIS_QUOTA_ERROR,
+    ErrorMessages.ODIS_RATE_LIMIT_ERROR,
     ErrorMessages.ODIS_AUTH_ERROR,
     ErrorMessages.ODIS_INPUT_ERROR,
     ErrorMessages.ODIS_CLIENT_ERROR,
@@ -158,15 +152,20 @@ export async function queryOdis<ResponseType>(
 
   return selectiveRetryAsyncWithBackOff(
     async () => {
-      const res = await fetch(odisUrl + endpoint, {
-        method: 'POST',
-        headers: {
-          Accept: 'application/json',
-          'Content-Type': 'application/json',
-          Authorization: authHeader,
-        },
-        body: bodyString,
-      })
+      let res: Response
+      try {
+        res = await fetch(odisUrl + endpoint, {
+          method: 'POST',
+          headers: {
+            Accept: 'application/json',
+            'Content-Type': 'application/json',
+            Authorization: signature,
+          },
+          body: bodyString,
+        })
+      } catch (error) {
+        throw new Error(`${ErrorMessages.ODIS_FETCH_ERROR}: ${error}`)
+      }
 
       if (res.ok) {
         debug('Response ok. Parsing.')
@@ -179,6 +178,89 @@ export async function queryOdis<ResponseType>(
       switch (res.status) {
         case 403:
           throw new Error(ErrorMessages.ODIS_QUOTA_ERROR)
+        case 429:
+          throw new Error(ErrorMessages.ODIS_RATE_LIMIT_ERROR)
+        case 400:
+          throw new Error(ErrorMessages.ODIS_INPUT_ERROR)
+        case 401:
+          throw new Error(ErrorMessages.ODIS_AUTH_ERROR)
+        default:
+          if (res.status >= 400 && res.status < 500) {
+            // Don't retry error codes in 400s
+            throw new Error(`${ErrorMessages.ODIS_CLIENT_ERROR} ${res.status}`)
+          }
+          throw new Error(`Unknown failure ${res.status}`)
+      }
+    },
+    3,
+    dontRetry,
+    []
+  )
+}
+
+/**
+ * Send the given domain request to ODIS (e.g. to get a POPRF evaluation or check quota).
+ *
+ * @param body Request to send in the body of the HTTP request.
+ * @param context Contains service URL and public to determine which instance to contact.
+ * @param endpoint Endpoint to query (e.g. '/domain/sign', '/domain/quotaStatus').
+ * @param responseSchema io-ts type for the expected response type. Provided to ensure type safety.
+ */
+export async function sendOdisDomainRequest<RequestType extends DomainRequest<Domain>>(
+  body: RequestType,
+  context: ServiceContext,
+  endpoint: DomainEndpoint,
+  responseSchema: t.Type<DomainResponse<RequestType>>
+): Promise<DomainResponse<RequestType>> {
+  debug(`Posting to ${endpoint}`)
+
+  const bodyString = JSON.stringify(body)
+
+  const { odisUrl } = context
+
+  const dontRetry = [
+    ErrorMessages.ODIS_QUOTA_ERROR,
+    ErrorMessages.ODIS_RATE_LIMIT_ERROR,
+    ErrorMessages.ODIS_AUTH_ERROR,
+    ErrorMessages.ODIS_INPUT_ERROR,
+    ErrorMessages.ODIS_CLIENT_ERROR,
+  ]
+
+  return selectiveRetryAsyncWithBackOff(
+    async () => {
+      let res: Response
+      try {
+        res = await fetch(odisUrl + endpoint, {
+          method: 'POST',
+          headers: {
+            Accept: 'application/json',
+            'Content-Type': 'application/json',
+          },
+          body: bodyString,
+        })
+      } catch (error) {
+        throw new Error(`${ErrorMessages.ODIS_FETCH_ERROR}: ${error}`)
+      }
+
+      if (res.ok) {
+        debug('Response ok. Parsing.')
+        const response = await res.json()
+
+        // Verify that the response is the type we expected, then return it.
+        const decoding = responseSchema.decode(response)
+        if (isLeft(decoding)) {
+          throw new Error(ErrorMessages.ODIS_RESPONSE_ERROR)
+        }
+        return decoding.right
+      }
+
+      debug(`Response not okay. Status ${res.status}`)
+
+      switch (res.status) {
+        case 403:
+          throw new Error(ErrorMessages.ODIS_QUOTA_ERROR)
+        case 429:
+          throw new Error(ErrorMessages.ODIS_RATE_LIMIT_ERROR)
         case 400:
           throw new Error(ErrorMessages.ODIS_INPUT_ERROR)
         case 401: