+++ packages/sdk/CHANGELOG.md
@@ -5,18 +5,32 @@ This package will follow the release process outlined [here](https://docs.celo.o
 
 
 ## Development (not published yet)
-### **[1.2.3--dev]**
+### **[1.5.3-dev]**
 Features
--  one-line summary - [#](link PR)
+- Adds the @celo/encrypted-backup package - [#8896] [#9348] [#9413] [#9434]
+
+[#8896]: https://github.com/celo-org/celo-monorepo/pull/8896
+[#9348]: https://github.com/celo-org/celo-monorepo/pull/9348
+[#9413]: https://github.com/celo-org/celo-monorepo/pull/9413
+[#9434]: https://github.com/celo-org/celo-monorepo/pull/9434
 
 Bug Fixes
 -  one-line summary - [#](link PR)
 
 Other Changes
 -  one-line summary - [#](link PR)
-
+-  migrated @celo/wallet-walletconnect from celo/celo-monorepo to celo/use-contractkit
 
 ## Published
+
+### **[1.5.0]**
+Features
+- Adds cREAL support to contract kit [#9040](https://github.com/celo-org/celo-monorepo/pull/9040)
+
+### **[1.4.2]**
+Bug Fixes
+-  Fixes wallet connect issue in @celo/wallet-walletconnect
+
 
 ### **[1.2.2]** -- 2021-07-08
 Features